Refactor components and remove unused files
- Removed Sidebar component and its associated logic.
- Deleted WhiteboardCanvas component and its related files.
- Cleaned up TextNote, Watch, and Weather components by removing unused imports.
- Extracted ComponentShelf and DraggableWhiteboardComponent into separate files.
- Implemented a new structure for managing components and their categories.
- Updated whiteboard logic to improve drag-and-drop functionality.
- Removed unused utility functions and hooks.
- Created a detailed refactoring plan to enhance code readability and maintainability.
